User feature implemented;
When the user copies the quotation to the clipboard, they want to discard the decimals for each component listed in the pasted
quotation.

The added feature checks if the string contains a notation '.' and if so, strips the string from there to remove the decimals.
The decimals are still included in the calculations, this only affects the clipboard memory quotation.